use u64 since bindgen stopped convert size_t to usize

+++ src/error.rs
@@ -2,8 +2,8 @@ macro_rules! check_result {
   ($condition: expr, $block: block) => {
     let errnum = $condition;
     if errnum < 0 {
-      let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE as usize];
-      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE);
+      let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE];
+      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE as u64);
       $block;
       return Err(tools::to_string(data.as_ptr()));
     }
@@ -11,8 +11,8 @@ macro_rules! check_result {
   ($condition: expr) => {
     let errnum = $condition;
     if errnum < 0 {
-      let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE as usize];
-      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE);
+      let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE];
+      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE as u64);
       return Err(tools::to_string(data.as_ptr()));
     }
   };
