+++ src/video_encoder.rs
@@ -129,8 +129,8 @@ impl VideoEncoder {
       let ret = avcodec_receive_packet(self.codec_context, packet.packet as *mut _);
 
       if ret == AVERROR(EAGAIN) || ret == AVERROR_EOF {
-        let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE as usize];
-        av_strerror(ret, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE);
+        let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE];
+        av_strerror(ret, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE as u64);
         trace!("{}", tools::to_string(data.as_ptr()));
         return Ok(false);
       }
